+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Latest
 
+- Adding support for ATLAS tag offset in plotting python API + Small fix for ratio uncertainty in histogram plots [!490](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/490)
 - Adding support for multiple signal classes [!414](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/414)